% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_conjugate.R
\name{.brsm_parse_prior_sd}
\alias{.brsm_parse_prior_sd}
\title{Parse Prior Scale/SD from String}
\usage{
.brsm_parse_prior_sd(prior_str)
}
\description{
Parse Prior Scale/SD from String
}
\keyword{internal}
