% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrix_to_long_format}
\alias{matrix_to_long_format}
\title{Convert Prediction Matrix to Long-Format Data Frame}
\usage{
matrix_to_long_format(pred_matrix, draw_ids, point_ids)
}
\arguments{
\item{pred_matrix}{A matrix of predictions (n_draws x n_points)}

\item{draw_ids}{Integer vector of draw identifiers}

\item{point_ids}{Integer vector of point identifiers}
}
\value{
A data frame with columns: draw, point_id, estimate
}
\description{
Convert Prediction Matrix to Long-Format Data Frame
}
\keyword{internal}
