% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{brsm-package}
\alias{brsm}
\alias{brsm-package}
\title{brsm: Bayesian Response Surface Methods}
\description{
Tools for analyzing posterior distributions of quadratic response surfaces from Bayesian model fits. Provides functions for computing stationary points, ridge analysis, steepest ascent paths, and visualization of response surfaces with full posterior uncertainty quantification.
}
\author{
\strong{Maintainer}: Arian Alai \email{aalai4@nebraska.edu}

}
